@@ -3,9 +3,9 @@
 # Table name: statistics
 #
 #  id             :integer          not null, primary key
-#  total_messages :integer          default(0)
-#  total_outgoing :integer          default(0)
-#  total_incoming :integer          default(0)
+#  total_messages :bigint(8)        default(0)
+#  total_outgoing :bigint(8)        default(0)
+#  total_incoming :bigint(8)        default(0)
 #
 
 class Statistic < ApplicationRecord
