@@ -15,11 +15,11 @@ module Postal
     def self.request(method, url, options = {})
       options[:headers] ||= {}
       uri = URI.parse(url)
-      request = method.new(uri.path.length == 0 ? "/" : uri.path)
+      request = method.new((uri.path.length == 0 ? "/" : uri.path) + (uri.query ? "?" + uri.query : ""))
       options[:headers].each { |k,v| request.add_field k, v }
 
-      if options[:username]
-        request.basic_auth(options[:username], options[:password])
+      if options[:username] || uri.user
+        request.basic_auth(options[:username] || uri.user, options[:password] || uri.password)
       end
 
       if options[:params].is_a?(Hash)
