@@ -5,8 +5,16 @@ module Postal
   module RabbitMQ
 
     def self.create_connection
+      bunny_host = [ 'localhost' ]
+
+      if Postal.config.rabbitmq&.host.is_a?(Array)
+        bunny_host = Postal.config.rabbitmq&.host
+      elsif Postal.config.rabbitmq&.host.is_a?(String)
+        bunny_host = [ Postal.config.rabbitmq&.host ]
+      end
+
       conn = Bunny.new(
-        :host => Postal.config.rabbitmq&.host || 'localhost',
+        :hosts => bunny_host,
         :port => Postal.config.rabbitmq&.port || 5672,
         :username => Postal.config.rabbitmq&.username || 'guest',
         :password => Postal.config.rabbitmq&.password || 'guest',
