@@ -39,7 +39,8 @@ module Postal
           ssl_context.cert = Postal.smtp_certificates[0]
           ssl_context.extra_chain_cert = Postal.smtp_certificates[1..-1]
           ssl_context.key  = Postal.smtp_private_key
-          ssl_context.ssl_version = "SSLv23"
+          ssl_context.ssl_version = Postal.config.smtp_server.ssl_version if Postal.config.smtp_server.ssl_version
+          ssl_context.ciphers = Postal.config.smtp_server.tls_ciphers if Postal.config.smtp_server.tls_ciphers
           ssl_context
         end
       end
@@ -142,7 +143,7 @@ module Postal
                   else
                     buffers[io] << io.readpartial(10240)
                   end
-                rescue EOFError, Errno::ECONNRESET
+                rescue EOFError, Errno::ECONNRESET, Errno::ETIMEDOUT
                   # Client went away
                   eof = true
                 end
